+++ cmd/frontend/db/repos.go
@@ -450,22 +450,22 @@ func parseIncludePattern(pattern string) (exact, like []string, regexp string, e
 // allMatchingStrings returns a complete list of the strings that re
 // matches, if it's possible to determine the list.
 func allMatchingStrings(re *regexpsyntax.Regexp) (exact, contains, prefix, suffix []string, err error) {
-	prog, err := regexpsyntax.Compile(re)
-	if err != nil {
-		return nil, nil, nil, nil, err
-	}
-
-	switch {
-	case re.Op == regexpsyntax.OpEmptyMatch:
+	switch re.Op {
+	case regexpsyntax.OpEmptyMatch:
 		return []string{""}, nil, nil, nil, nil
-	case re.Op == regexpsyntax.OpLiteral:
+	case regexpsyntax.OpLiteral:
+		prog, err := regexpsyntax.Compile(re)
+		if err != nil {
+			return nil, nil, nil, nil, err
+		}
+
 		prefix, complete := prog.Prefix()
 		if complete {
 			return nil, []string{prefix}, nil, nil, nil
 		}
 		return nil, nil, nil, nil, nil
 
-	case re.Op == regexpsyntax.OpCharClass:
+	case regexpsyntax.OpCharClass:
 		// Only handle simple case of one range.
 		if len(re.Rune) == 2 {
 			len := int(re.Rune[1] - re.Rune[0] + 1)
@@ -482,16 +482,21 @@ func allMatchingStrings(re *regexpsyntax.Regexp) (exact, contains, prefix, suffi
 		}
 		return nil, nil, nil, nil, nil
 
-	case re.Op == regexpsyntax.OpBeginText:
+	case regexpsyntax.OpStar:
+		if len(re.Sub) == 1 && (re.Sub[0].Op == regexpsyntax.OpAnyCharNotNL || re.Sub[0].Op == regexpsyntax.OpAnyChar) {
+			return nil, []string{""}, nil, nil, nil
+		}
+
+	case regexpsyntax.OpBeginText:
 		return nil, nil, []string{""}, nil, nil
 
-	case re.Op == regexpsyntax.OpEndText:
+	case regexpsyntax.OpEndText:
 		return nil, nil, nil, []string{""}, nil
 
-	case re.Op == regexpsyntax.OpCapture:
+	case regexpsyntax.OpCapture:
 		return allMatchingStrings(re.Sub0[0])
 
-	case re.Op == regexpsyntax.OpConcat:
+	case regexpsyntax.OpConcat:
 		var begin, end bool
 		for i, sub := range re.Sub {
 			if sub.Op == regexpsyntax.OpBeginText && i == 0 {
@@ -542,7 +547,7 @@ func allMatchingStrings(re *regexpsyntax.Regexp) (exact, contains, prefix, suffi
 		}
 		return nil, exact, nil, nil, nil
 
-	case re.Op == regexpsyntax.OpAlternate:
+	case regexpsyntax.OpAlternate:
 		for _, sub := range re.Sub {
 			subexact, subcontains, subprefix, subsuffix, err := allMatchingStrings(sub)
 			if err != nil {