+++ web/src/global/GlobalAlerts.tsx
@@ -73,7 +73,7 @@ export class GlobalAlerts extends React.PureComponent<Props, State> {
                                 className="global-alerts__alert"
                             />
                         )}
-                        {/* Only show if the user has already enabled repositories; if not yet, the user wouldn't experience any Docker for Mac perf issues anyway. */}
+                        {/* Only show if the user has already added repositories; if not yet, the user wouldn't experience any Docker for Mac perf issues anyway. */}
                         {window.context.likelyDockerOnMac && <DockerForMacAlert className="global-alerts__alert" />}
                         {this.state.siteFlags.alerts.map((alert, i) => (
                             <GlobalAlert key={i} alert={alert} className="global-alerts__alert" />