+++ web/src/repo/settings/RepoSettingsOptionsPage.tsx
@@ -4,16 +4,13 @@ import * as React from 'react'
 import { RouteComponentProps } from 'react-router'
 import { Subject, Subscription } from 'rxjs'
 import { switchMap } from 'rxjs/operators'
-import { REPO_DELETE_CONFIRMATION_MESSAGE } from '.'
 import * as GQL from '../../../../shared/src/graphql/schema'
 import { ExternalServiceCard } from '../../components/ExternalServiceCard'
 import { Form } from '../../components/Form'
 import { PageTitle } from '../../components/PageTitle'
-import { deleteRepository, setRepositoryEnabled } from '../../site-admin/backend'
 import { getExternalService } from '../../site-admin/externalServices'
 import { eventLogger } from '../../tracking/eventLogger'
 import { fetchRepository } from './backend'
-import { ActionContainer } from './components/ActionContainer'
 
 interface Props extends RouteComponentProps<any> {
     repo: GQL.IRepository
@@ -106,56 +103,7 @@ export class RepoSettingsOptionsPage extends React.PureComponent<Props, State> {
                         />
                     </div>
                 </Form>
-                {services.length === 0 && [
-                    <ActionContainer
-                        title={this.state.repo.enabled ? 'Disable access' : 'Enable access'}
-                        description={
-                            this.state.repo.enabled
-                                ? 'Disable access to the repository to prevent users from searching and browsing the repository.'
-                                : 'The repository is disabled. Enable it to allow users to search and view the repository.'
-                        }
-                        buttonClassName={this.state.repo.enabled ? 'btn-danger' : 'btn-success'}
-                        buttonLabel={this.state.repo.enabled ? 'Disable access' : 'Enable access'}
-                        flashText="Updated"
-                        run={this.state.repo.enabled ? this.disableRepository : this.enableRepository}
-                    />,
-                    <ActionContainer
-                        title="Delete repository"
-                        description="Permanently removes this repository and all associated data from Sourcegraph. The original repository on the code host is not affected. If this repository was added by a configured code host, then it will be re-added during the next sync."
-                        buttonClassName="btn-danger"
-                        buttonLabel="Delete this repository"
-                        run={this.deleteRepository}
-                    />,
-                ]}
             </div>
         )
-    }
-
-    private enableRepository = () =>
-        setRepositoryEnabled(this.state.repo.id, true)
-            .toPromise()
-            .then(() => {
-                this.repoUpdates.next()
-                this.props.onDidUpdateRepository({ enabled: true })
-            })
-    private disableRepository = () =>
-        setRepositoryEnabled(this.state.repo.id, false)
-            .toPromise()
-            .then(() => {
-                this.repoUpdates.next()
-                this.props.onDidUpdateRepository({ enabled: false })
-            })
-
-    private deleteRepository = () => {
-        if (!window.confirm(REPO_DELETE_CONFIRMATION_MESSAGE)) {
-            return Promise.resolve()
-        }
-
-        return deleteRepository(this.state.repo.id)
-            .toPromise()
-            .then(() => {
-                this.repoUpdates.next()
-                this.props.history.push('/site-admin/repositories')
-            })
     }
 }